Remove unavailable items from `depends` field of library.properties
The depends field of the library.properties metadata file specifies the dependencies that should be installed along with
the library by the Arduino Library Manager.

This field must contain only the names of libraries that are available for installation via Library Manager.

The presence of any items which are not in Library Manager causes installation of the library to fail:

- Arduino IDE 1.x: "no protocol:" error
- Arduino IDE 2.x: fails silently
- Arduino CLI: "No valid dependencies solution found" error

The removed dependencies are "platform bundled libraries", which are included with the installation of each Arduino boards platform, and thus are managed by Arduino Boards Manager, not Library Manager. For this reason, there is no reason for these items to be listed in the depends field.